+++ R/desc.R
@@ -22,21 +22,57 @@ fill_desc <- function(
   author_email, 
   repo_url = NULL,
   pkg = get_golem_wd()
-  ){
+){
+  
   path <- normalizePath(pkg)
+  
   desc <- desc::description$new(
     file = file.path(path, "DESCRIPTION")
   )
-  desc$set("Authors@R", glue("person('{author_first_name}', '{author_last_name}', email = '{author_email}', role = c('cre', 'aut'))"))
-  desc$del("Maintainer")
-  desc$set_version("0.0.0.9000")
-  desc$set(Package = pkg_name)
-  desc$set(Title = pkg_title)
-  desc$set(Description = pkg_description)
-  if_not_null(repo_url, desc$set("URL", repo_url))
-  if_not_null(repo_url, desc$set("BugReports", glue("{repo_url}/issues")))
-  desc$write(file = "DESCRIPTION")
-  cat_bullet("DESCRIPTION file modified", bullet = "tick", bullet_col = "green")
+  desc$set(
+    "Authors@R", 
+    glue("person('{author_first_name}', '{author_last_name}', email = '{author_email}', role = c('cre', 'aut'))")
+  )
+  desc$del(
+    keys = "Maintainer"
+  )
+  desc$set_version(
+    version = "0.0.0.9000"
+  )
+  desc$set(
+    Package = pkg_name
+  )
+  desc$set(
+    Title = pkg_title
+  )
+  desc$set(
+    Description = pkg_description
+  )
+  if_not_null(
+    repo_url, 
+    desc$set(
+      "URL", 
+      repo_url
+    )
+  )
+  if_not_null(
+    repo_url, 
+    desc$set(
+      "BugReports", 
+      glue("{repo_url}/issues"
+      )
+    )
+  )
+  
+  desc$write(
+    file = "DESCRIPTION"
+  )
+  
+  cat_bullet(
+    "DESCRIPTION file modified", 
+    bullet = "tick", 
+    bullet_col = "green"
+  )
 }